corrections to explanation of gradient f(x,y)
I think it is important to emphasise that grad f is a vector since we're dealing with 'direction' (vectors are a literal description of direction).  I tried to simplify the explanation as much as possible.  The previous explanation of grad f was incorrect - it is not a proportion and we do not divide the partial derivatives to find the slope.